add blueprints to database schema and models
