+++ nexus/db-model/src/inventory.rs
@@ -4,6 +4,7 @@
 
 //! Types for representing the hardware/software inventory in the database
 
+use crate::omicron_zone_config::{OmicronZone, OmicronZoneNic};
 use crate::schema::{
     hw_baseboard_id, inv_caboose, inv_collection, inv_collection_error,
     inv_omicron_zone, inv_omicron_zone_nic, inv_root_of_trust,
@@ -14,8 +15,7 @@ use crate::{
     impl_enum_type, ipv6, ByteCount, Generation, MacAddr, Name, SqlU16, SqlU32,
     SqlU8,
 };
-use anyhow::{anyhow, ensure};
-use anyhow::{bail, Context};
+use anyhow::anyhow;
 use chrono::DateTime;
 use chrono::Utc;
 use diesel::backend::Backend;
@@ -26,10 +26,8 @@ use diesel::serialize::ToSql;
 use diesel::{serialize, sql_types};
 use ipnetwork::IpNetwork;
 use nexus_types::inventory::{
-    BaseboardId, Caboose, Collection, OmicronZoneType, PowerState, RotPage,
-    RotSlot,
+    BaseboardId, Caboose, Collection, PowerState, RotPage, RotSlot,
 };
-use std::net::SocketAddrV6;
 use uuid::Uuid;
 
 // See [`nexus_types::inventory::PowerState`].
@@ -750,165 +748,29 @@ impl InvOmicronZone {
         sled_id: Uuid,
         zone: &nexus_types::inventory::OmicronZoneConfig,
     ) -> Result<InvOmicronZone, anyhow::Error> {
-        let id = zone.id;
-        let underlay_address = ipv6::Ipv6Addr::from(zone.underlay_address);
-        let mut nic_id = None;
-        let mut dns_gz_address = None;
-        let mut dns_gz_address_index = None;
-        let mut ntp_ntp_servers = None;
-        let mut ntp_dns_servers = None;
-        let mut ntp_ntp_domain = None;
-        let mut nexus_external_tls = None;
-        let mut nexus_external_dns_servers = None;
-        let mut snat_ip = None;
-        let mut snat_first_port = None;
-        let mut snat_last_port = None;
-        let mut second_service_ip = None;
-        let mut second_service_port = None;
-
-        let (zone_type, primary_service_sockaddr_str, dataset) = match &zone
-            .zone_type
-        {
-            OmicronZoneType::BoundaryNtp {
-                address,
-                ntp_servers,
-                dns_servers,
-                domain,
-                nic,
-                snat_cfg,
-            } => {
-                ntp_ntp_servers = Some(ntp_servers.clone());
-                ntp_dns_servers = Some(dns_servers.clone());
-                ntp_ntp_domain = domain.clone();
-                snat_ip = Some(IpNetwork::from(snat_cfg.ip));
-                snat_first_port = Some(SqlU16::from(snat_cfg.first_port));
-                snat_last_port = Some(SqlU16::from(snat_cfg.last_port));
-                nic_id = Some(nic.id);
-                (ZoneType::BoundaryNtp, address, None)
-            }
-            OmicronZoneType::Clickhouse { address, dataset } => {
-                (ZoneType::Clickhouse, address, Some(dataset))
-            }
-            OmicronZoneType::ClickhouseKeeper { address, dataset } => {
-                (ZoneType::ClickhouseKeeper, address, Some(dataset))
-            }
-            OmicronZoneType::CockroachDb { address, dataset } => {
-                (ZoneType::CockroachDb, address, Some(dataset))
-            }
-            OmicronZoneType::Crucible { address, dataset } => {
-                (ZoneType::Crucible, address, Some(dataset))
-            }
-            OmicronZoneType::CruciblePantry { address } => {
-                (ZoneType::CruciblePantry, address, None)
-            }
-            OmicronZoneType::ExternalDns {
-                dataset,
-                http_address,
-                dns_address,
-                nic,
-            } => {
-                nic_id = Some(nic.id);
-                let sockaddr = dns_address
-                    .parse::<std::net::SocketAddr>()
-                    .with_context(|| {
-                        format!(
-                            "parsing address for external DNS server {:?}",
-                            dns_address
-                        )
-                    })?;
-                second_service_ip = Some(sockaddr.ip());
-                second_service_port = Some(SqlU16::from(sockaddr.port()));
-                (ZoneType::ExternalDns, http_address, Some(dataset))
-            }
-            OmicronZoneType::InternalDns {
-                dataset,
-                http_address,
-                dns_address,
-                gz_address,
-                gz_address_index,
-            } => {
-                dns_gz_address = Some(ipv6::Ipv6Addr::from(gz_address));
-                dns_gz_address_index = Some(SqlU32::from(*gz_address_index));
-                let sockaddr = dns_address
-                    .parse::<std::net::SocketAddr>()
-                    .with_context(|| {
-                        format!(
-                            "parsing address for internal DNS server {:?}",
-                            dns_address
-                        )
-                    })?;
-                second_service_ip = Some(sockaddr.ip());
-                second_service_port = Some(SqlU16::from(sockaddr.port()));
-                (ZoneType::InternalDns, http_address, Some(dataset))
-            }
-            OmicronZoneType::InternalNtp {
-                address,
-                ntp_servers,
-                dns_servers,
-                domain,
-            } => {
-                ntp_ntp_servers = Some(ntp_servers.clone());
-                ntp_dns_servers = Some(dns_servers.clone());
-                ntp_ntp_domain = domain.clone();
-                (ZoneType::InternalNtp, address, None)
-            }
-            OmicronZoneType::Nexus {
-                internal_address,
-                external_ip,
-                nic,
-                external_tls,
-                external_dns_servers,
-            } => {
-                nic_id = Some(nic.id);
-                nexus_external_tls = Some(*external_tls);
-                nexus_external_dns_servers = Some(external_dns_servers.clone());
-                second_service_ip = Some(*external_ip);
-                (ZoneType::Nexus, internal_address, None)
-            }
-            OmicronZoneType::Oximeter { address } => {
-                (ZoneType::Oximeter, address, None)
-            }
-        };
-
-        let dataset_zpool_name =
-            dataset.map(|d| d.pool_name.as_str().to_string());
-        let primary_service_sockaddr = primary_service_sockaddr_str
-            .parse::<std::net::SocketAddrV6>()
-            .with_context(|| {
-                format!(
-                    "parsing socket address for primary IP {:?}",
-                    primary_service_sockaddr_str
-                )
-            })?;
-        let (primary_service_ip, primary_service_port) = (
-            ipv6::Ipv6Addr::from(*primary_service_sockaddr.ip()),
-            SqlU16::from(primary_service_sockaddr.port()),
-        );
-
-        Ok(InvOmicronZone {
+        let zone = OmicronZone::new(sled_id, zone)?;
+        Ok(Self {
             inv_collection_id,
-            sled_id,
-            id,
-            underlay_address,
-            zone_type,
-            primary_service_ip,
-            primary_service_port,
-            second_service_ip: second_service_ip.map(IpNetwork::from),
-            second_service_port,
-            dataset_zpool_name,
-            nic_id,
-            dns_gz_address,
-            dns_gz_address_index,
-            ntp_ntp_servers,
-            ntp_dns_servers: ntp_dns_servers
-                .map(|list| list.into_iter().map(IpNetwork::from).collect()),
-            ntp_domain: ntp_ntp_domain,
-            nexus_external_tls,
-            nexus_external_dns_servers: nexus_external_dns_servers
-                .map(|list| list.into_iter().map(IpNetwork::from).collect()),
-            snat_ip,
-            snat_first_port,
-            snat_last_port,
+            sled_id: zone.sled_id,
+            id: zone.id,
+            underlay_address: zone.underlay_address,
+            zone_type: zone.zone_type,
+            primary_service_ip: zone.primary_service_ip,
+            primary_service_port: zone.primary_service_port,
+            second_service_ip: zone.second_service_ip,
+            second_service_port: zone.second_service_port,
+            dataset_zpool_name: zone.dataset_zpool_name,
+            nic_id: zone.nic_id,
+            dns_gz_address: zone.dns_gz_address,
+            dns_gz_address_index: zone.dns_gz_address_index,
+            ntp_ntp_servers: zone.ntp_ntp_servers,
+            ntp_dns_servers: zone.ntp_dns_servers,
+            ntp_domain: zone.ntp_domain,
+            nexus_external_tls: zone.nexus_external_tls,
+            nexus_external_dns_servers: zone.nexus_external_dns_servers,
+            snat_ip: zone.snat_ip,
+            snat_first_port: zone.snat_first_port,
+            snat_last_port: zone.snat_last_port,
         })
     }
 
@@ -916,169 +778,29 @@ impl InvOmicronZone {
         self,
         nic_row: Option<InvOmicronZoneNic>,
     ) -> Result<nexus_types::inventory::OmicronZoneConfig, anyhow::Error> {
-        let address = SocketAddrV6::new(
-            std::net::Ipv6Addr::from(self.primary_service_ip),
-            *self.primary_service_port,
-            0,
-            0,
-        )
-        .to_string();
-
-        // Assemble a value that we can use to extract the NIC _if necessary_
-        // and report an error if it was needed but not found.
-        //
-        // Any error here should be impossible.  By the time we get here, the
-        // caller should have provided `nic_row` iff there's a corresponding
-        // `nic_id` in this row, and the ids should match up.  And whoever
-        // created this row ought to have provided a nic_id iff this type of
-        // zone needs a NIC.  This last issue is not under our control, though,
-        // so we definitely want to handle that as an operational error.  The
-        // others could arguably be programmer errors (i.e., we could `assert`),
-        // but it seems excessive to crash here.
-        //
-        // Note that we immediately return for any of the caller errors here.
-        // For the other error, we will return only later, if some code path
-        // below tries to use `nic` when it's not present.
-        let nic = match (self.nic_id, nic_row) {
-            (Some(expected_id), Some(nic_row)) => {
-                ensure!(expected_id == nic_row.id, "caller provided wrong NIC");
-                Ok(nic_row.into_network_interface_for_zone(self.id)?)
-            }
-            (None, None) => Err(anyhow!(
-                "expected zone to have an associated NIC, but it doesn't"
-            )),
-            (Some(_), None) => bail!("caller provided no NIC"),
-            (None, Some(_)) => bail!("caller unexpectedly provided a NIC"),
-        };
-
-        // Similarly, assemble a value that we can use to extract the dataset,
-        // if necessary.  We only return this error if code below tries to use
-        // this value.
-        let dataset = self
-            .dataset_zpool_name
-            .map(|zpool_name| -> Result<_, anyhow::Error> {
-                Ok(nexus_types::inventory::OmicronZoneDataset {
-                    pool_name: zpool_name.parse().map_err(|e| {
-                        anyhow!("parsing zpool name {:?}: {}", zpool_name, e)
-                    })?,
-                })
-            })
-            .transpose()?
-            .ok_or_else(|| anyhow!("expected dataset zpool name, found none"));
-
-        // Do the same for the DNS server address.
-        let dns_address =
-            match (self.second_service_ip, self.second_service_port) {
-                (Some(dns_ip), Some(dns_port)) => {
-                    Ok(std::net::SocketAddr::new(dns_ip.ip(), *dns_port)
-                        .to_string())
-                }
-                _ => Err(anyhow!(
-                    "expected second service IP and port, \
-                            found one missing"
-                )),
-            };
-
-        // Do the same for NTP zone properties.
-        let ntp_dns_servers = self
-            .ntp_dns_servers
-            .ok_or_else(|| anyhow!("expected list of DNS servers, found null"))
-            .map(|list| {
-                list.into_iter().map(|ipnetwork| ipnetwork.ip()).collect()
-            });
-        let ntp_ntp_servers =
-            self.ntp_ntp_servers.ok_or_else(|| anyhow!("expected ntp_servers"));
-
-        let zone_type = match self.zone_type {
-            ZoneType::BoundaryNtp => {
-                let snat_cfg = match (
-                    self.snat_ip,
-                    self.snat_first_port,
-                    self.snat_last_port,
-                ) {
-                    (Some(ip), Some(first_port), Some(last_port)) => {
-                        nexus_types::inventory::SourceNatConfig {
-                            ip: ip.ip(),
-                            first_port: *first_port,
-                            last_port: *last_port,
-                        }
-                    }
-                    _ => bail!(
-                        "expected non-NULL snat properties, \
-                        found at least one NULL"
-                    ),
-                };
-                OmicronZoneType::BoundaryNtp {
-                    address,
-                    dns_servers: ntp_dns_servers?,
-                    domain: self.ntp_domain,
-                    nic: nic?,
-                    ntp_servers: ntp_ntp_servers?,
-                    snat_cfg,
-                }
-            }
-            ZoneType::Clickhouse => {
-                OmicronZoneType::Clickhouse { address, dataset: dataset? }
-            }
-            ZoneType::ClickhouseKeeper => {
-                OmicronZoneType::ClickhouseKeeper { address, dataset: dataset? }
-            }
-            ZoneType::CockroachDb => {
-                OmicronZoneType::CockroachDb { address, dataset: dataset? }
-            }
-            ZoneType::Crucible => {
-                OmicronZoneType::Crucible { address, dataset: dataset? }
-            }
-            ZoneType::CruciblePantry => {
-                OmicronZoneType::CruciblePantry { address }
-            }
-            ZoneType::ExternalDns => OmicronZoneType::ExternalDns {
-                dataset: dataset?,
-                dns_address: dns_address?,
-                http_address: address,
-                nic: nic?,
-            },
-            ZoneType::InternalDns => OmicronZoneType::InternalDns {
-                dataset: dataset?,
-                dns_address: dns_address?,
-                http_address: address,
-                gz_address: *self.dns_gz_address.ok_or_else(|| {
-                    anyhow!("expected dns_gz_address, found none")
-                })?,
-                gz_address_index: *self.dns_gz_address_index.ok_or_else(
-                    || anyhow!("expected dns_gz_address_index, found none"),
-                )?,
-            },
-            ZoneType::InternalNtp => OmicronZoneType::InternalNtp {
-                address,
-                dns_servers: ntp_dns_servers?,
-                domain: self.ntp_domain,
-                ntp_servers: ntp_ntp_servers?,
-            },
-            ZoneType::Nexus => OmicronZoneType::Nexus {
-                internal_address: address,
-                nic: nic?,
-                external_tls: self
-                    .nexus_external_tls
-                    .ok_or_else(|| anyhow!("expected 'external_tls'"))?,
-                external_ip: self
-                    .second_service_ip
-                    .ok_or_else(|| anyhow!("expected second service IP"))?
-                    .ip(),
-                external_dns_servers: self
-                    .nexus_external_dns_servers
-                    .ok_or_else(|| anyhow!("expected 'external_dns_servers'"))?
-                    .into_iter()
-                    .map(|i| i.ip())
-                    .collect(),
-            },
-            ZoneType::Oximeter => OmicronZoneType::Oximeter { address },
-        };
-        Ok(nexus_types::inventory::OmicronZoneConfig {
+        let zone = OmicronZone {
+            sled_id: self.sled_id,
             id: self.id,
-            underlay_address: std::net::Ipv6Addr::from(self.underlay_address),
-            zone_type,
-        })
+            underlay_address: self.underlay_address,
+            zone_type: self.zone_type,
+            primary_service_ip: self.primary_service_ip,
+            primary_service_port: self.primary_service_port,
+            second_service_ip: self.second_service_ip,
+            second_service_port: self.second_service_port,
+            dataset_zpool_name: self.dataset_zpool_name,
+            nic_id: self.nic_id,
+            dns_gz_address: self.dns_gz_address,
+            dns_gz_address_index: self.dns_gz_address_index,
+            ntp_ntp_servers: self.ntp_ntp_servers,
+            ntp_dns_servers: self.ntp_dns_servers,
+            ntp_domain: self.ntp_domain,
+            nexus_external_tls: self.nexus_external_tls,
+            nexus_external_dns_servers: self.nexus_external_dns_servers,
+            snat_ip: self.snat_ip,
+            snat_first_port: self.snat_first_port,
+            snat_last_port: self.snat_last_port,
+        };
+        zone.into_omicron_zone_config(nic_row.map(OmicronZoneNic::from))
     }
 }
 
@@ -1096,63 +818,45 @@ pub struct InvOmicronZoneNic {
     slot: SqlU8,
 }
 
+impl From<InvOmicronZoneNic> for OmicronZoneNic {
+    fn from(value: InvOmicronZoneNic) -> Self {
+        OmicronZoneNic {
+            id: value.id,
+            name: value.name,
+            ip: value.ip,
+            mac: value.mac,
+            subnet: value.subnet,
+            vni: value.vni,
+            is_primary: value.is_primary,
+            slot: value.slot,
+        }
+    }
+}
+
 impl InvOmicronZoneNic {
     pub fn new(
         inv_collection_id: Uuid,
         zone: &nexus_types::inventory::OmicronZoneConfig,
     ) -> Result<Option<InvOmicronZoneNic>, anyhow::Error> {
-        match &zone.zone_type {
-            OmicronZoneType::ExternalDns { nic, .. }
-            | OmicronZoneType::BoundaryNtp { nic, .. }
-            | OmicronZoneType::Nexus { nic, .. } => {
-                // We do not bother storing the NIC's kind and associated id
-                // because it should be inferrable from the other information
-                // that we have.  Verify that here.
-                ensure!(
-                    matches!(
-                        nic.kind,
-                        nexus_types::inventory::NetworkInterfaceKind::Service(
-                            id
-                        ) if id == zone.id
-                    ),
-                    "expected zone's NIC kind to be \"service\" and the \
-                    id to match the zone's id ({})",
-                    zone.id
-                );
-
-                Ok(Some(InvOmicronZoneNic {
-                    inv_collection_id,
-                    id: nic.id,
-                    name: Name::from(nic.name.clone()),
-                    ip: IpNetwork::from(nic.ip),
-                    mac: MacAddr::from(nic.mac),
-                    subnet: IpNetwork::from(nic.subnet.clone()),
-                    vni: SqlU32::from(u32::from(nic.vni)),
-                    is_primary: nic.primary,
-                    slot: SqlU8::from(nic.slot),
-                }))
-            }
-            _ => Ok(None),
-        }
+        let zone_nic = OmicronZoneNic::new(zone)?;
+        Ok(zone_nic.map(|nic| Self {
+            inv_collection_id,
+            id: nic.id,
+            name: nic.name,
+            ip: nic.ip,
+            mac: nic.mac,
+            subnet: nic.subnet,
+            vni: nic.vni,
+            is_primary: nic.is_primary,
+            slot: nic.slot,
+        }))
     }
 
     pub fn into_network_interface_for_zone(
         self,
         zone_id: Uuid,
     ) -> Result<nexus_types::inventory::NetworkInterface, anyhow::Error> {
-        Ok(nexus_types::inventory::NetworkInterface {
-            id: self.id,
-            ip: self.ip.ip(),
-            kind: nexus_types::inventory::NetworkInterfaceKind::Service(
-                zone_id,
-            ),
-            mac: *self.mac,
-            name: self.name.into(),
-            primary: self.is_primary,
-            slot: *self.slot,
-            vni: omicron_common::api::external::Vni::try_from(*self.vni)
-                .context("parsing VNI")?,
-            subnet: self.subnet.into(),
-        })
+        let zone_nic = OmicronZoneNic::from(self);
+        zone_nic.into_network_interface_for_zone(zone_id)
     }
 }
